TS Linting
Adds editor support to JS files so you can still get type checking for your source code as well. This doesn't change any output with the `.d.ts` files, nor the need to move to TS directly.

This initial version is solely without the strict flags enabled. There are still errors that it is not fond of though, the additional flags can help make it more strict as well though. JSDoc to the rescue!